Added migration, renamed Id and moved datamodels
